src/main/java/rfl/spring/recipe/converters/NotesToNotesCommand.java
package rfl.spring.recipe.converters;

import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;
import rfl.spring.recipe.command.NotesCommand;
import rfl.spring.recipe.domain.Notes;

@Component
public class NotesToNotesCommand implements Converter<Notes, NotesCommand> {

    @Synchronized
    @Nullable
    @Override
    public NotesCommand convert(Notes notes) {
        if (notes == null) {
            return null;
        }

        NotesCommand notesc = new NotesCommand();
        notesc.setId(notes.getId());
        notesc.setRecipeNotes(notes.getRecipeNotes());
        return notesc;
    }
}
